Use page template as cache key for it's pages
This removes touching the section when updating pages.
Instead sections and items touches their page template when updated and
hence should the page template be used when caching collections of
pages.
