+++ app/models/item.rb
@@ -23,9 +23,7 @@ class Item
   key :restricted, Boolean
 
   key :published_on, Time
-  
-  key :association_ids, Array, typecast: 'ObjectId'
-  
+
   timestamps!
 
   taggable
@@ -48,9 +46,7 @@ class Item
   before_create :set_created_by
   before_save :set_updated_by
   before_save :sort_data
-  before_save :store_association_ids
-  
-  after_save :touch_associations
+
   after_save proc { |page| page.delay.update_tank_indexes }
   after_destroy proc { |page| page.delete_tank_indexes }
 
@@ -114,29 +110,6 @@ private
 
   def set_updated_by
     self.updated_by = User.current
-  end
-  
-  def store_association_ids
-    self.association_ids = find_association_ids
-  end
-
-  def touch_associations
-    Item.set({ association_ids: self.id }, { updated_at: self.updated_at.utc })
-  end
-
-  def find_association_ids(source = nil)
-    collection = source || self.data
-    result = []
-
-    collection.find_all { |d| d.active? }.each do |d|
-      if d.respond_to?(:page_id)
-        result << d.page_id
-      elsif d.respond_to?(:data) && d.data.any?
-        result += find_association_ids(d.data)
-      end
-    end
-
-    result.compact.uniq
   end
 
 end
