+++ app/models/page.rb
@@ -18,7 +18,6 @@ class Page < Section
   key :position, Integer
 
   before_create :move_to_list_bottom
-  after_save :touch_section
 
   scope :with_page_template, lambda { |page_template_id|
     where(:page_template_id => page_template_id)
@@ -165,7 +164,4 @@ private
     index_node && (index_node.restricted? || index_node.ancestors.detect { |n| n.restricted? })
   end
 
-  def touch_section
-    Section.set({ page_template_id: self.page_template_id }, { updated_at: self.updated_at.utc })
-  end
 end
